[thread-spawn] Mark additions with their criticality to the proposal
After many discussions on the topic, including [this one], there seems
to be early consensus on needing to do something along the lines of this
proposal but there are differing views on the specific form this might
take. I fully expect the proposal to change with more discussion and
especially as we prototype and troubleshoot the idea. To capture this in
the proposal, I added text indicating how critical each addition is:
e.g., if we removed `shared` attributes, this proposal should be closed
and replaced by another one. This level of criticality varies for each
addition.

[this one]: https://github.com/abrown/thread-spawn/discussions/1