Add Thread-Bound Data to the documented JS API
As discussed in #30 and #37, add a JS API that allows arbitrary objects to be
wrapped and converted to shared externrefs with dynamic checks ensuring the
wrapped object is only ever observed on the original thread. Refer to the
existing FAQ entry for discussion on how the lifetimes of the wrapper and
wrapped object might interact.

Resolves #37.