Describe WAT parsing in more detail
This change expands how we explain the WebAssembly text format in the
overview based on some comments in #61 re: the ordering of `shared` and
implementation experience in `wasm-tools` trying to both write spec
tests and parse them. Some of this could be pieced together in the
original overview but several points were a bit too ambiguous for me.
Closes #61.