Update memory ordering encodings
Add separate `release` and `acquire` encodings since an `acquire-release` fence
is not the same as an `acquire` fence. Encode all orderings as `u8` instead of
`u32` and in the case of memory instructions, use a bit in the memarg to signal
the presence of a separate ordering immediate rather than encoding the ordering
directly in the memarg. For RMW ops, pack both a read ordering and a write
ordering into the single `u8` ordering immediate.

This design is similar to the one described in
https://github.com/WebAssembly/shared-everything-threads/issues/60#issuecomment-2136475267,
but uses `release` and `acquire` instead of `release-acquire` for load and store
orderings, using `release-acquire` only for fences.

Fixes #59.
Fixes #60.