Update memory ordering encodings (#67)
Encode all orderings as `u8` instead of `u32` and use a bit in memarg to
signal the presence of a separate ordering immediate rather than
encoding orderings directly in memarg. For RMW ops, pack both a read
ordering and a write ordering into the single `u8` ordering immediate.

This design is similar to the one described in
https://github.com/WebAssembly/shared-everything-threads/issues/60#issuecomment-2136475267.

Fixes #59.
Fixes #60.