Change from type variables to bot-share for polymorphism (#88)
Using type variables to allow instructions to be polymorphic over the
sharedness of their inputs allows for principle typing (up to type
variables), but it does not allow for forward principle typing, since
there is no most-precise instantiation of the type variable to choose
when the available input is a bottom stack. For example, this code
sequence could produce either a `(ref any)` or a `(ref (shared any))`:

```
(unreachable) (any.convert_extern)
```

To fix this, instead introduce a bottom shareability, `bot-share`, that
is analogous to the bottom heap type and used only for validation.

Fixes #80.