Upstream initial test suite from `wasm-tools` (#85)
This change starts this proposal's test suite by adding encoding,
decoding, and validation tests for the new instructions. This is a
wholesale copy of the current state of the `wasm-tools` [tests]. Some
things to note:
- some of the tests were generated mechanically to attempt to "cover all
possibilities;" not all proposals do this, though, so it's unclear
whether we want to here (it's verbose, but hopefully more complete)
- these tests include the component model built-ins, which are not
executable in many engines; it's unclear whether to keep that file (the
proposal speaks of this) or not
- these tests are a start, not the end state: we're missing anything to
do with "thread locals," `pause`, and possibly some other bits.

[tests]:
https://github.com/bytecodealliance/wasm-tools/tree/5a83828/tests/local/shared-everything-threads

---------

Co-authored-by: Thomas Lively <tlively123@gmail.com>